+++ src/funding/venmo/config.jsx
@@ -6,9 +6,9 @@ import {
   VenmoLogoInlineSVG,
   LOGO_COLOR,
 } from "@paypal/sdk-logos/src";
-import { PLATFORM } from "@paypal/sdk-constants/src";
+import { DISPLAY_ONLY_VALUES, PLATFORM } from "@paypal/sdk-constants/src";
 
-import { BUTTON_COLOR, BUTTON_LAYOUT } from "../../constants";
+import { BUTTON_COLOR, BUTTON_LAYOUT, BUTTON_FLOW } from "../../constants";
 import { DEFAULT_FUNDING_CONFIG, type FundingSourceConfig } from "../common";
 
 import { WalletLabel, Label } from "./template";
@@ -17,9 +17,11 @@ export function getVenmoConfig(): FundingSourceConfig {
   return {
     ...DEFAULT_FUNDING_CONFIG,
 
+    flows: [BUTTON_FLOW.PURCHASE, BUTTON_FLOW.VAULT_WITHOUT_PURCHASE],
+
     layouts: [BUTTON_LAYOUT.HORIZONTAL, BUTTON_LAYOUT.VERTICAL],
 
-    eligible: ({ experiment, shippingChange }) => {
+    eligible: ({ experiment, shippingChange, displayOnly, flow }) => {
       if (experiment && experiment.enableVenmo === false) {
         return false;
       }
@@ -28,6 +30,20 @@ export function getVenmoConfig(): FundingSourceConfig {
         experiment &&
         experiment.venmoWebEnabled === false &&
         shippingChange
+      ) {
+        return false;
+      }
+
+      if (
+        shippingChange &&
+        displayOnly?.includes(DISPLAY_ONLY_VALUES.VAULTABLE)
+      ) {
+        return false;
+      }
+
+      if (
+        flow === BUTTON_FLOW.VAULT_WITHOUT_PURCHASE &&
+        experiment?.venmoVaultWithoutPurchase !== true
       ) {
         return false;
       }