+++ src/hosted-buttons/utils.js
@@ -2,6 +2,7 @@
 
 import { request, memoize } from "@krakenjs/belter/src";
 import {
+  getLogger,
   buildDPoPHeaders,
   getSDKHost,
   getClientID,
@@ -9,6 +10,10 @@ import {
   getMerchantID as getSDKMerchantID,
 } from "@paypal/sdk-client/src";
 import { FUNDING } from "@paypal/sdk-constants/src";
+import { SUPPORTED_FUNDING_SOURCES } from "@paypal/funding-components/src";
+import { ZalgoPromise } from "@krakenjs/zalgo-promise/src";
+
+import { getButtonsComponent, type ButtonsComponent } from "../zoid/buttons";
 
 import type {
   ButtonVariables,
@@ -18,6 +23,17 @@ import type {
   HostedButtonDetailsParams,
   OnApprove,
   RenderForm,
+  GetFlexDirectionArgs,
+  GetFlexDirection,
+  Color,
+  FundingSources,
+  ApplyButtonStylesProps,
+  HostedButtonPreferences,
+  NcpResponsePreferences,
+  ButtonPreferences,
+  GetButtonsProps,
+  RenderStandaloneButtonProps,
+  RenderDefaultButtonProps,
 } from "./types";
 
 const entryPoint = "SDK";
@@ -31,11 +47,13 @@ const getHeaders = (accessToken?: string) => ({
 });
 
 export const getMerchantID = (): string | void => {
-  // The SDK supports mutiple merchant IDs, but hosted buttons only
-  // have one merchant id as a query parameter to the SDK script.
+  // The SDK supports Multi-Seller Payments (MSP, i.e sending multiple merchant IDs), but hosted buttons
+  // does not support this. Only one merchant id can be passed as a query parameter to the SDK script
   // https://github.com/paypal/paypal-sdk-client/blob/c58e35f8f7adbab76523eb25b9c10543449d2d29/src/script.js#L144
+  // https://developer.paypal.com/docs/multiparty/checkout/multiseller-payments/
   const merchantIds = getSDKMerchantID();
   if (merchantIds.length > 1) {
+    getLogger().error("ncps_multiple_merchant_ids", { merchantIds });
     throw new Error("Multiple merchant-ids are not supported.");
   }
   return merchantIds[0];
@@ -72,6 +90,41 @@ export const createAccessToken: CreateAccessToken = memoize<CreateAccessToken>(
   }
 );
 
+export const getButtonPreferences = ({
+  button_preferences: buttonPreferences,
+  eligible_funding_methods: eligibleFundingMethods,
+}: NcpResponsePreferences): HostedButtonPreferences => {
+  if (!buttonPreferences?.length || !eligibleFundingMethods?.length) {
+    const preferences = {
+      buttonPreferences,
+      eligibleFundingMethods,
+    };
+
+    getLogger().error("ncps_missing_preferences", { preferences });
+
+    throw new Error(
+      `Expected preferences to be populated, received: ${JSON.stringify({
+        preferences,
+      })}`
+    );
+  }
+
+  return {
+    // Remove any buttons that are not included in eligibleFundingMethods.
+    // If the funding method is "default", we want to keep it in the preferences, and decide which
+    // button should be rendered in its place in renderStandaloneButton()
+    buttonPreferences: buttonPreferences.filter(
+      (fundingMethod) =>
+        eligibleFundingMethods.includes(fundingMethod) ||
+        fundingMethod === "default"
+    ),
+    // Sort the eligible funding methods returned from /ncp/api/form-fields in the order that they would appear in the smart stack.
+    eligibleFundingMethods: SUPPORTED_FUNDING_SOURCES.filter((fundingMethod) =>
+      eligibleFundingMethods.includes(fundingMethod)
+    ),
+  };
+};
+
 const getButtonVariable = (variables: ButtonVariables, key: string): string =>
   variables?.find((variable) => variable.name === key)?.value ?? "";
 
@@ -88,18 +141,41 @@ export const getHostedButtonDetails: HostedButtonDetailsParams = async ({
 
   // $FlowIssue request returns ZalgoPromise
   const { body } = response;
-  const variables = body.button_details.link_variables;
+  const {
+    link_variables: variables,
+    js_sdk_container_id: buttonContainerId,
+    preferences,
+  } = body.button_details;
+
+  const shouldIncludePreferences = preferences && body.version === "2";
+
   return {
     style: {
       layout: getButtonVariable(variables, "layout"),
       shape: getButtonVariable(variables, "shape"),
       color: getButtonVariable(variables, "color"),
       label: getButtonVariable(variables, "button_text"),
+      tagline: getButtonVariable(variables, "tagline") === "true",
+      height: parseInt(getButtonVariable(variables, "height"), 10) || undefined,
     },
+    enableDPoP: getButtonVariable(variables, "enable_dpop") === "true",
+    version: body.version,
+    buttonContainerId: buttonContainerId || "spb-container",
     html: body.html,
     htmlScript: body.html_script,
+    ...(shouldIncludePreferences && {
+      preferences: getButtonPreferences(preferences),
+    }),
   };
 };
+
+export function getElementFromSelector(
+  selector: string | HTMLElement
+): HTMLElement | null {
+  return typeof selector === "string"
+    ? document.querySelector(selector)
+    : selector;
+}
 
 /**
  * Attaches form fields (html) to the given selector, and
@@ -111,8 +187,7 @@ export const renderForm: RenderForm = ({
   htmlScript,
   selector,
 }) => {
-  const elm =
-    typeof selector === "string" ? document.querySelector(selector) : selector;
+  const elm = getElementFromSelector(selector);
   if (elm) {
     elm.innerHTML = html + htmlScript;
     const newScriptEl = document.createElement("script");
@@ -228,4 +303,151 @@ export const buildHostedButtonOnApprove = ({
       return response;
     });
   };
+};
+
+export const getFlexDirection = ({
+  layout,
+}: GetFlexDirectionArgs): GetFlexDirection => ({
+  flexDirection: layout === "horizontal" ? "row" : "column",
+});
+
+export const getButtonColor = (
+  color: Color,
+  fundingSource: FundingSources
+): Color => {
+  const colorMap = {
+    gold: {
+      paypal: "gold",
+      venmo: "blue",
+      paylater: "gold",
+    },
+    blue: {
+      paypal: "blue",
+      venmo: "silver",
+      paylater: "blue",
+    },
+    black: {
+      paypal: "black",
+      venmo: "black",
+      paylater: "black",
+    },
+    white: {
+      paypal: "white",
+      venmo: "white",
+      paylater: "white",
+    },
+    silver: {
+      paypal: "silver",
+      venmo: "blue",
+      paylater: "silver",
+    },
+  };
+
+  return colorMap[color][fundingSource];
+};
+
+export const applyContainerStyles = ({
+  flexDirection,
+  buttonContainerId,
+}: ApplyButtonStylesProps): void => {
+  const buttonContainer = document.querySelector(`#${buttonContainerId}`);
+
+  if (!buttonContainer) {
+    getLogger().error("ncps_button_container_missing", {
+      buttonContainerId,
+    });
+
+    throw new Error(`Element with id ${buttonContainerId} not found.`);
+  }
+
+  buttonContainer.style.flexDirection = flexDirection;
+};
+
+/**
+ * Filters out all eligible funding methods that are already specified in button preferences
+ */
+export const getDefaultButtonOptions = ({
+  buttonPreferences,
+  eligibleFundingMethods,
+}: HostedButtonPreferences): ButtonPreferences => {
+  return eligibleFundingMethods.filter(
+    (fundingSource: string) => !buttonPreferences.includes(fundingSource)
+  );
+};
+
+/**
+ * Gets buttons component instance.
+ */
+export const getButtons = ({
+  fundingSource,
+  buttonOptions,
+}: GetButtonsProps): ButtonsComponent => {
+  const Buttons = getButtonsComponent();
+
+  const { style } = buttonOptions;
+
+  // $FlowFixMe
+  return Buttons({
+    ...buttonOptions,
+    fundingSource,
+    style: {
+      ...style,
+      // $FlowFixMe
+      color: getButtonColor(style.color, fundingSource),
+    },
+  });
+};
+
+/**
+ * Handles logic for each specified button preference.
+ */
+export const renderStandaloneButton = ({
+  fundingSource,
+  buttonContainerId,
+  buttonOptions,
+}: RenderStandaloneButtonProps): ZalgoPromise<void> | void => {
+  const standaloneButton = getButtons({
+    fundingSource,
+    buttonOptions,
+  });
+
+  // $FlowFixMe
+  if (standaloneButton.isEligible()) {
+    // $FlowFixMe
+    return standaloneButton.render(`#${buttonContainerId}`);
+  }
+
+  getLogger().error(`ncps_standalone_${fundingSource}_ineligible`);
+};
+
+/**
+ * Handles logic for "default" button preference.
+ */
+export const renderDefaultButton = ({
+  eligibleDefaultButtons,
+  buttonContainerId,
+  buttonOptions,
+}: RenderDefaultButtonProps): void => {
+  const eligibleButtons = [...eligibleDefaultButtons];
+
+  // If we exhaust all default options, we don't render any button.
+  while (eligibleButtons.length) {
+    const fundingSource = eligibleButtons[0];
+
+    const standaloneButton = getButtons({
+      fundingSource,
+      buttonOptions,
+    });
+
+    // If the funding source is eligible, render button & return to end loop.
+    // $FlowFixMe
+    if (standaloneButton.isEligible()) {
+      // $FlowFixMe
+      return standaloneButton.render(`#${buttonContainerId}`);
+    }
+
+    // If funding source is ineligible, log error and move to next funding option.
+    getLogger().error(`ncps_standalone_${fundingSource}_ineligible`);
+    eligibleButtons.shift();
+  }
 };