+++ src/zoid/buttons/component.jsx
@@ -19,6 +19,7 @@ import {
   getCSPNonce,
   getBuyerCountry,
   getClientAccessToken,
+  getCustomerId,
   getPlatform,
   getPartnerAttributionID,
   getCorrelationID,
@@ -38,6 +39,7 @@ import {
   getVersion,
   getDisableSetCookie,
   getExperimentation,
+  getFirstRenderExperiments,
   getSDKAttribute,
   getJsSdkLibrary,
 } from "@paypal/sdk-client/src";
@@ -54,7 +56,6 @@ import {
   isApplePaySupported,
   supportsPopups as userAgentSupportsPopups,
   noop,
-  isLocalStorageEnabled,
 } from "@krakenjs/belter/src";
 import {
   FUNDING,
@@ -79,6 +80,7 @@ import {
   type ButtonProps,
 } from "../../ui/buttons/props";
 import { isFundingEligible } from "../../funding";
+import { CLASS } from "../../constants";
 
 import { containerTemplate } from "./container";
 import { PrerenderedButtons } from "./prerender";
@@ -106,6 +108,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
     autoResize: {
       width: false,
       height: true,
+      element: `.${CLASS.AUTORESIZE_CONTAINER}`,
     },
 
     containerTemplate,
@@ -120,18 +123,6 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
 
     prerenderTemplate: ({ state, props, doc, event }) => {
       const { buttonSessionID } = props;
-
-      if (!isLocalStorageEnabled()) {
-        getLogger()
-          .info("localstorage_inaccessible_possible_private_browsing")
-          .track({
-            [FPTI_KEY.BUTTON_SESSION_UID]: buttonSessionID,
-            [FPTI_KEY.CONTEXT_TYPE]: "button_session_id",
-            [FPTI_KEY.CONTEXT_ID]: buttonSessionID,
-            [FPTI_KEY.TRANSITION]:
-              "localstorage_inaccessible_possible_private_browsing",
-          });
-      }
 
       event.on(EVENT.PRERENDERED, () => {
         // CPL stands for Consumer Perceived Latency
@@ -148,17 +139,13 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
             [FPTI_KEY.PAGE]: "main:xo:paypal-components:smart-payment-buttons",
             [FPTI_KEY.CPL_COMP_METRICS]: JSON.stringify(cplPhases?.comp || {}),
           };
-          getLogger()
-            .info("CPL_LATENCY_METRICS_FIRST_RENDER")
-            .track(cplLatencyMetrics);
+          getLogger().track(cplLatencyMetrics);
         } catch (err) {
-          getLogger()
-            .info("button_render_CPL_instrumentation_log_error")
-            .track({
-              err: err.message || "CPL_LOG_PHASE_ERROR",
-              details: err.details,
-              stack: JSON.stringify(err.stack || err),
-            });
+          getLogger().track({
+            err: err.message || "CPL_LOG_PHASE_ERROR",
+            details: err.details,
+            stack: JSON.stringify(err.stack || err),
+          });
         }
       });
 
@@ -201,6 +188,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         createBillingAgreement,
         createSubscription,
         createVaultSetupToken,
+        displayOnly,
       } = props;
 
       const flow = determineFlow({
@@ -243,6 +231,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
           supportsPopups,
           supportedNativeBrowser,
           experiment,
+          displayOnly,
         })
       ) {
         return {
@@ -334,6 +323,13 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         required: false,
         queryParam: true,
         value: getClientAccessToken,
+      },
+
+      customerId: {
+        type: "string",
+        required: false,
+        queryParam: true,
+        value: getCustomerId,
       },
 
       clientID: {
@@ -462,12 +458,13 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
       experiment: {
         type: "object",
         queryParam: true,
-        value: () => {
-          const experiments = getButtonExperiments();
-          return experiments;
-        },
+        value: () => ({
+          ...getButtonExperiments(),
+          ...getFirstRenderExperiments(),
+        }),
       },
-
+      // TODO first-render-experiment-cleanup
+      // verify if this is needed/used now that were putting the first render experiments in experiment param above
       experimentation: {
         type: "object",
         queryParam: true,
@@ -520,6 +517,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
             createBillingAgreement,
             createSubscription,
             createVaultSetupToken,
+            displayOnly,
           } = props;
 
           const flow = determineFlow({
@@ -548,6 +546,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
               applePaySupport,
               supportsPopups,
               supportedNativeBrowser,
+              displayOnly,
             })
           ) {
             throw new Error(`${fundingSource} is not eligible`);
@@ -717,7 +716,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
             offerCountryCode,
             creditProductIdentifier,
           }) => {
-            const { message, clientID, merchantID, currency, buttonSessionID } =
+            const { message, clientID, currency, buttonSessionID, merchantID } =
               props;
             const amount = message?.amount;
 
@@ -730,7 +729,6 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
                 [FPTI_KEY.CONTEXT_ID]: buttonSessionID,
                 [FPTI_KEY.CONTEXT_TYPE]: "button_session_id",
                 [FPTI_KEY.EVENT_NAME]: "message_click",
-                [FPTI_KEY.SELLER_ID]: merchantID?.join(","),
                 [FPTI_KEY.BUTTON_MESSAGE_OFFER_TYPE]: offerType,
                 [FPTI_KEY.BUTTON_MESSAGE_CREDIT_PRODUCT_IDENTIFIER]:
                   creditProductIdentifier,
@@ -741,7 +739,8 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
                 [FPTI_KEY.BUTTON_MESSAGE_OFFER_COUNTRY]: offerCountryCode,
                 [FPTI_KEY.BUTTON_MESSAGE_CURRENCY]: currency,
                 [FPTI_KEY.BUTTON_MESSAGE_AMOUNT]: amount,
-              });
+              })
+              .flush();
 
             const modalInstance = await getModal(clientID, merchantID);
             return modalInstance?.show({
@@ -758,10 +757,10 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         required: false,
         value: ({ props }) => {
           return () => {
+            const { clientID, merchantID } = props;
             // offerType, messageType, offerCountryCode, and creditProductIdentifier are passed in and may be used in an upcoming message hover logging feature
 
             // lazy loads the modal, to be memoized and executed onMessageClick
-            const { clientID, merchantID } = props;
             return getModal(clientID, merchantID);
           };
         },
@@ -776,8 +775,17 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
             messageType,
             offerCountryCode,
             creditProductIdentifier,
+            merchantID: serverMerchantId,
           }) => {
-            const { message, buttonSessionID, currency, merchantID } = props;
+            // merchantID that comes from props is an array
+            const { message, buttonSessionID, currency } = props;
+
+            // override with server id if partner does not exist
+            if (serverMerchantId) {
+              getLogger().addTrackingBuilder(() => ({
+                [FPTI_KEY.SELLER_ID]: serverMerchantId,
+              }));
+            }
 
             getLogger()
               .info("button_message_render")
@@ -788,7 +796,6 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
                 [FPTI_KEY.CONTEXT_ID]: buttonSessionID,
                 [FPTI_KEY.CONTEXT_TYPE]: "button_session_id",
                 [FPTI_KEY.EVENT_NAME]: "message_render",
-                [FPTI_KEY.SELLER_ID]: merchantID?.join(","),
                 [FPTI_KEY.BUTTON_MESSAGE_OFFER_TYPE]: offerType,
                 [FPTI_KEY.BUTTON_MESSAGE_CREDIT_PRODUCT_IDENTIFIER]:
                   creditProductIdentifier,
@@ -799,7 +806,8 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
                 [FPTI_KEY.BUTTON_MESSAGE_CURRENCY]: currency,
                 [FPTI_KEY.BUTTON_MESSAGE_OFFER_COUNTRY]: offerCountryCode,
                 [FPTI_KEY.BUTTON_MESSAGE_AMOUNT]: message?.amount,
-              });
+              })
+              .flush();
           };
         },
       },
@@ -821,6 +829,19 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
       onShippingOptionsChange: {
         type: "function",
         required: false,
+      },
+
+      hasShippingCallback: {
+        type: "boolean",
+        required: false,
+        queryParam: true,
+        value: ({ props }) => {
+          return Boolean(
+            props.onShippingChange ||
+              props.onShippingAddressChange ||
+              props.onShippingOptionsChange
+          );
+        },
       },
 
       pageType: {